+++ components/ad2mqtt/ad2mqtt.cpp
@@ -53,10 +53,10 @@ static const char *TAG = "MQTT";
 #define SK_PREFILTER_REGEX   "P"
 #define SK_OPEN_REGEX_LIST   "O"
 #define SK_CLOSED_REGEX_LIST "C"
-#define SK_FAULT_REGEX_LIST  "F"
+#define SK_TROUBLE_REGEX_LIST  "F"
 #define SK_OPEN_OUTPUT_FMT   "o"
 #define SK_CLOSED_OUTPUT_FMT "c"
-#define SK_FAULT_OUTPUT_FMT  "f"
+#define SK_TROUBLE_OUTPUT_FMT  "f"
 
 #define MQTT_TOPIC_PREFIX "ad2iot"
 #define MQTT_LWT_TOPIC_SUFFIX "will"
@@ -370,7 +370,7 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
 
             case SK_OPEN_REGEX_LIST[0]: // Open state REGEX list editor
             case SK_CLOSED_REGEX_LIST[0]: // Closed state REGEX list editor
-            case SK_FAULT_REGEX_LIST[0]: // Fault state REGEX list editor
+            case SK_TROUBLE_REGEX_LIST[0]: // Trouble state REGEX list editor
                 // Add index to file name to track N number of elements.
                 ad2_copy_nth_arg(arg1, instring, 4);
                 i = std::atoi (arg1.c_str());
@@ -385,8 +385,8 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
                     if (buf[0] == SK_CLOSED_REGEX_LIST[0]) {
                         tmpsz = "CLOSED";
                     }
-                    if (buf[0] == SK_FAULT_REGEX_LIST[0]) {
-                        tmpsz = "FAULT";
+                    if (buf[0] == SK_TROUBLE_REGEX_LIST[0]) {
+                        tmpsz = "TROUBLE";
                     }
                     ad2_set_nv_slot_key_string(key.c_str(), slot, sk.c_str(), arg2.c_str());
                     ad2_printf_host("%s smartswitch #%i REGEX filter #%02i for state '%s' to '%s'.\r\n", op.c_str(), slot, i, tmpsz.c_str(), arg2.c_str());
@@ -397,7 +397,7 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
 
             case SK_OPEN_OUTPUT_FMT[0]: // Open state output format string
             case SK_CLOSED_OUTPUT_FMT[0]: // Closed state output format string
-            case SK_FAULT_OUTPUT_FMT[0]: // Fault state output format string
+            case SK_TROUBLE_OUTPUT_FMT[0]: // Trouble state output format string
                 // consume the arge and to EOL
                 ad2_copy_nth_arg(arg1, instring, 4, true);
                 if (buf[0] == SK_OPEN_OUTPUT_FMT[0]) {
@@ -406,8 +406,8 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
                 if (buf[0] == SK_CLOSED_OUTPUT_FMT[0]) {
                     tmpsz = "CLOSED";
                 }
-                if (buf[0] == SK_FAULT_OUTPUT_FMT[0]) {
-                    tmpsz = "FAULT";
+                if (buf[0] == SK_TROUBLE_OUTPUT_FMT[0]) {
+                    tmpsz = "TROUBLE";
                 }
                 ad2_set_nv_slot_key_string(key.c_str(), slot, sk.c_str(), arg1.c_str());
                 ad2_printf_host("Setting smartswitch #%i output format string for '%s' state to '%s'.\r\n", slot, tmpsz.c_str(), arg1.c_str());
@@ -428,10 +428,10 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
                                SK_PREFILTER_REGEX
                                SK_OPEN_REGEX_LIST
                                SK_CLOSED_REGEX_LIST
-                               SK_FAULT_REGEX_LIST
+                               SK_TROUBLE_REGEX_LIST
                                SK_OPEN_OUTPUT_FMT
                                SK_CLOSED_OUTPUT_FMT
-                               SK_FAULT_OUTPUT_FMT;
+                               SK_TROUBLE_OUTPUT_FMT;
 
             ad2_printf_host("MQTT SmartSwitch #%i report\r\n", slot);
             // sub key suffix.
@@ -477,7 +477,7 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
                     break;
                 case SK_OPEN_REGEX_LIST[0]:
                 case SK_CLOSED_REGEX_LIST[0]:
-                case SK_FAULT_REGEX_LIST[0]:
+                case SK_TROUBLE_REGEX_LIST[0]:
                     // * Find all sub keys and show info.
                     if (c == SK_OPEN_REGEX_LIST[0]) {
                         tmpsz = "OPEN";
@@ -485,8 +485,8 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
                     if (c == SK_CLOSED_REGEX_LIST[0]) {
                         tmpsz = "CLOSED";
                     }
-                    if (c == SK_FAULT_REGEX_LIST[0]) {
-                        tmpsz = "FAULT";
+                    if (c == SK_TROUBLE_REGEX_LIST[0]) {
+                        tmpsz = "TROUBLE";
                     }
                     for ( i = 1; i < MAX_SEARCH_KEYS; i++ ) {
                         out = "";
@@ -500,15 +500,15 @@ static void _cli_cmd_mqtt_smart_alert_switch(std::string &subcmd, char *instring
                     break;
                 case SK_OPEN_OUTPUT_FMT[0]:
                 case SK_CLOSED_OUTPUT_FMT[0]:
-                case SK_FAULT_OUTPUT_FMT[0]:
+                case SK_TROUBLE_OUTPUT_FMT[0]:
                     if (c == SK_OPEN_OUTPUT_FMT[0]) {
                         tmpsz = "OPEN";
                     }
                     if (c == SK_CLOSED_OUTPUT_FMT[0]) {
                         tmpsz = "CLOSED";
                     }
-                    if (c == SK_FAULT_OUTPUT_FMT[0]) {
-                        tmpsz = "FAULT";
+                    if (c == SK_TROUBLE_OUTPUT_FMT[0]) {
+                        tmpsz = "TROUBLE";
                     }
                     out = "";
                     ad2_get_nv_slot_key_string(key.c_str(), slot, sk.c_str(), out);
@@ -633,12 +633,12 @@ static struct cli_command mqtt_cmd_list[] = {
         "      - [C] Close(OFF) state regex search string list management.\r\n"
         "        - {arg1}: Index # 1-8\r\n"
         "        - {arg2}: Regex string for this slot or empty string to clear\r\n"
-        "      - [F] Fault state regex search string list management.\r\n"
+        "      - [F] Trouble state regex search string list management.\r\n"
         "        - {arg1}: Index # 1-8\r\n"
         "        - {arg2}: Regex string for this slot or empty  string to clear\r\n"
         "      - [o] Open output format string.\r\n"
         "      - [c] Close output format string.\r\n"
-        "      - [f] Fault output format string.\r\n\r\n", _cli_cmd_mqtt_command_router
+        "      - [f] Trouble output format string.\r\n\r\n", _cli_cmd_mqtt_command_router
     },
 };
 
@@ -721,7 +721,7 @@ void mqtt_init()
     AD2Parse.subscribeTo(ON_ARM, mqtt_on_state_change, (void *)ON_ARM);
     AD2Parse.subscribeTo(ON_DISARM, mqtt_on_state_change, (void *)ON_DISARM);
     AD2Parse.subscribeTo(ON_CHIME_CHANGE, mqtt_on_state_change, (void *)ON_CHIME_CHANGE);
-    AD2Parse.subscribeTo(ON_FIRE, mqtt_on_state_change, (void *)ON_FIRE);
+    AD2Parse.subscribeTo(ON_FIRE_CHANGE, mqtt_on_state_change, (void *)ON_FIRE_CHANGE);
     AD2Parse.subscribeTo(ON_POWER_CHANGE, mqtt_on_state_change, (void *)ON_POWER_CHANGE);
     AD2Parse.subscribeTo(ON_READY_CHANGE, mqtt_on_state_change, (void *)ON_READY_CHANGE);
     AD2Parse.subscribeTo(ON_LOW_BATTERY, mqtt_on_state_change, (void *)ON_LOW_BATTERY);
@@ -744,10 +744,10 @@ void mqtt_init()
         // We at least need some output format or skip
         ad2_get_nv_slot_key_string(key.c_str(), i, SK_OPEN_OUTPUT_FMT, es1->OPEN_OUTPUT_FORMAT);
         ad2_get_nv_slot_key_string(key.c_str(), i, SK_CLOSED_OUTPUT_FMT, es1->CLOSED_OUTPUT_FORMAT);
-        ad2_get_nv_slot_key_string(key.c_str(), i, SK_FAULT_OUTPUT_FMT, es1->FAULT_OUTPUT_FORMAT);
+        ad2_get_nv_slot_key_string(key.c_str(), i, SK_TROUBLE_OUTPUT_FMT, es1->TROUBLE_OUTPUT_FORMAT);
         if ( es1->OPEN_OUTPUT_FORMAT.length()
                 || es1->CLOSED_OUTPUT_FORMAT.length()
-                || es1->FAULT_OUTPUT_FORMAT.length() ) {
+                || es1->TROUBLE_OUTPUT_FORMAT.length() ) {
 
             std::string notify_types_sz = "";
             std::vector<std::string> notify_types_v;
@@ -763,8 +763,8 @@ void mqtt_init()
             }
             ad2_get_nv_slot_key_string(key.c_str(), i, SK_PREFILTER_REGEX, es1->PRE_FILTER_REGEX);
 
-            // Load all regex search patterns for OPEN,CLOSE and FAULT sub keys.
-            std::string regex_sk_list = SK_FAULT_REGEX_LIST SK_CLOSED_REGEX_LIST SK_OPEN_REGEX_LIST;
+            // Load all regex search patterns for OPEN,CLOSE and TROUBLE sub keys.
+            std::string regex_sk_list = SK_TROUBLE_REGEX_LIST SK_CLOSED_REGEX_LIST SK_OPEN_REGEX_LIST;
             for(char& c : regex_sk_list) {
                 std::string sk = ad2_string_printf("%c", c);
                 for ( int a = 1; a < MAX_SEARCH_KEYS; a++) {
@@ -778,8 +778,8 @@ void mqtt_init()
                         if (c == SK_CLOSED_REGEX_LIST[0]) {
                             es1->CLOSED_REGEX_LIST.push_back(out);
                         }
-                        if (c == SK_FAULT_REGEX_LIST[0]) {
-                            es1->FAULT_REGEX_LIST.push_back(out);
+                        if (c == SK_TROUBLE_REGEX_LIST[0]) {
+                            es1->TROUBLE_REGEX_LIST.push_back(out);
                         }
                     }
                 }